requirements.txt
Django==4.2.29
djangorestframework